Refactoring to better reflect what is happening
 - Renamed "gateway" to persistence. That's really what we are doing.
 - Removed the interface references in the class names. It's redundant.
 - Created a placeholder for preference storage persistence in MySQL
