+++ lib/Wikia/src/Persistence/User/PreferenceStorageMySQL.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Wikia\Persistence\User;
+
+use Wikia\Service\User\PreferencePersistence;
+
+class PreferenceStorageMySQL implements PreferencePersistence {
+
+}
