+++ lib/Wikia/src/Service/User/PreferenceService.php
@@ -1,76 +1,10 @@
 <?php
-/**
- *
- * User preference service.
- * Assumptions:
- *	- The user exists (persisted in the database) and we have a valid user id.
- *	- Only the currently logged in user as set in the gateway WikiaUserId is
- *	  able to change their preferences
- *
- * Open questions:
- *	- How do we handle the preference white list on this side? Do we?
- *	- How do we standardize on a set of exceptions that will cross the libraries we use?
- *	  Things that could go wrong:
- *	  - gateway error e.g. 500 with API or database exception locally
- *	  - request error e.g. 400
- *	  - user not found 404
- *	  - unauthorized
- *	  - success 200
- *
- */
+
 namespace Wikia\Service\User;
 
-use Wikia\Domain\User\Preference;
+interface PreferenceService {
 
-class PreferenceService implements PreferenceServiceInterface {
+	public function setPreferences( $userId, $preferences );
+	public function getPreferences( $userId );
 
-	private $gateway;
-
-	function __construct( PreferenceGatewayInterface $gateway ) {
-		$this->gateway = $gateway;
-	}
-
-	public function setPreferences( $userId, $preferences ) {
-		if ( !is_array( $preferences ) || empty( $preferences ) ) {
-			return false;
-		}
-
-		$this->authenticateUser( $userId );
-
-		return $this->gateway->save( $userId, $preferences );
-	}
-
-	public function getPreferences( $userId ) {
-		$this->authenticateUser( $userId );
-
-		$result = $this->gateway->get( $userId );
-		if (!is_array($result)) {
-			return [];
-		}
-
-		$preferences = [];
-		try {
-		 $preferences = $this->gatewayResultToPreferenceArray( $result );
-		} catch ( \Wikia\Service\GatewayNotFoundException $e ) {
-		}
-
-		return $preferences;
-	}
-
-	protected function authenticateUser( $userId ) {
-		if ( $userId !== $this->gateway->getWikiaUserId() ) {
-			throw new \Wikia\Service\GatewayUnauthorizedException( "Unauthorized to set preferences." );
-		}
-	}
-
-	public function gatewayResultToPreferenceArray(array $result) {
-		$preferences = [];
-		foreach( $result as $index => $row ) {
-			if (isset($row["name"]) && isset($row["value"])) {
-				$preferences[] = new Preference($row["name"], $row["value"]);
-			}
-		}
-
-		return $preferences;
-	}
 }
