--- lib/Wikia/src/Service/User/PreferenceServiceInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Wikia\Service\User;
-
-interface PreferenceServiceInterface {
-
-	public function setPreferences( $userId, $preferences );
-	public function getPreferences( $userId );
-
-}
